Remove dependencies
