+++ RL2/trainer/ppo.py
@@ -5,7 +5,7 @@ import wandb
 
 from RL2.trainer import Trainer
 from RL2.datasets import RLDataset, get_dataloader
-from RL2.workers import Actor, Rollout, Critic
+from RL2.workers import Actor, Critic
 from RL2.utils.algorithms import (
     compute_approx_kl,
     compute_gae,
@@ -14,7 +14,7 @@ from RL2.utils.algorithms import (
 from RL2.utils.comm import initialize_global_process_group
 from RL2.utils.checkpointing import load_ckpt, save_ckpt, save_model
 from RL2.utils.logging import time_logger
-
+from RL2.utils.rollout_loader import load_rollout_class
 
 class PPOTrainer(Trainer):
 
@@ -30,7 +30,8 @@ class PPOTrainer(Trainer):
         if config.adv.estimator == "gae":
             self.critic = Critic(config.critic)
             self.critic.scheduler = self.prepare_scheduler(self.critic)
-        self.rollout = Rollout(config.rollout)    
+        rollout_class = load_rollout_class(config.rollout)
+        self.rollout = rollout_class(config.rollout)    
 
     def get_dataloader(self, train: bool):
