+++ RL2/workers/__init__.py
@@ -1,4 +1,4 @@
 from .base import Worker
 from .actor import Actor
-from .rollout_wrapper import Rollout  # Use wrapper for automatic selection
+from .rollout import Rollout
 from .critic import Critic
